+++ src/main/java/frc/robot/commands/Shooter/RampDownShooter.java
@@ -18,15 +18,17 @@ public class RampDownShooter extends CommandBase {
  private double mStartingVoltage;
  private boolean mFinished;
 
-  /** Creates a new RampDownShooter. */
   public RampDownShooter(Shooter pShooter, double pRampTime) {
     sShooter = pShooter;
     mRampTime = pRampTime;
+
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(sShooter);
   }
 
-  // Called when the command is initially scheduled.
+  /**
+   * 
+   */
   public void initialize() {
     mStartingVoltage = sShooter.GetMotorOutputPercent();
     mStartTime = Timer.getFPGATimestamp();
@@ -37,12 +39,14 @@ public class RampDownShooter extends CommandBase {
     SmartDashboard.putNumber("Ramp Start V", mStartingVoltage);
   }
 
-  // Called every time the scheduler runs while the command is scheduled.
+  /**
+   * 
+   */
   public void execute() {
-    if (mVoltage > 0){
+    if (mVoltage > 0) {
       mVoltage = -(mStartingVoltage / mRampTime) * (mCurrentTime - mStartTime) + mStartingVoltage;
       sShooter.RunShooter(mVoltage);
-    }else{
+    } else {
       mVoltage = 0;
       sShooter.RunShooter(mVoltage);
       mFinished = true;
@@ -52,12 +56,10 @@ public class RampDownShooter extends CommandBase {
     mCurrentTime = Timer.getFPGATimestamp();
   }
 
-  // Called once the command ends or is interrupted.
   public void end(boolean interrupted) {
     sShooter.RunShooter(0);
   }
 
-  // Returns true when the command should end.
   public boolean isFinished() {
     return mFinished;
   }
