+++ src/main/java/frc/robot/subsystems/Vision/USBVision.java
@@ -13,7 +13,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 public class USBVision extends SubsystemBase {
   
-  // Actual Camera Stream - Web Address of http://roboRIO-2509-FRC.local/1181/?action=stream
+  // Actual Camera Stream sent back from the robot
   public static UsbCamera camera = CameraServer.getInstance().startAutomaticCapture("BALL", 0);
 
   // 'Sinks' the camera stream into a processable image
@@ -24,7 +24,7 @@ public class USBVision extends SubsystemBase {
   public static CvSource outputStream = CameraServer.getInstance().putVideo("BALL2", 160, 120); // 160, 120
 
 
-  // Pipeline Declarlations
+  // Generated Pipeline from Grip application (Ctrl+G in Grip, select Java)
   ProofPipeline proof_pipeline = new ProofPipeline();
 
   /* Matrixes for Values of Images */
@@ -42,10 +42,10 @@ public class USBVision extends SubsystemBase {
   }
 
   public FoundContour[] getContours() {
+    
     FoundContour[] Contours = new FoundContour[5];
-    // for (FoundContour Contour : Contours) {
-    //   Contour = new FoundContour(-1, -1 , -1, false);
-    // }
+      
+    // Resets values in Countours array
     Contours[0] = new FoundContour(-1, -1 , -1, false);
     Contours[1] = new FoundContour(-1, -1 , -1, false);
     Contours[2] = new FoundContour(-1, -1 , -1, false);
@@ -61,28 +61,47 @@ public class USBVision extends SubsystemBase {
       return Contours;
     }
 
-    // Will show the processed image up to the
+    // DEBUGGING: Will show the processed image up to the
     // blur step
-    proof_pipeline.process(source);
-    outputStream.putFrame(proof_pipeline.blurOutput());
+    // proof_pipeline.process(source);
+    // outputStream.putFrame(proof_pipeline.blurOutput());
 
+
+    // Gets frame from source
     cvSink.grabFrame(source);
+
+    // If no countours are found (tests if output array in
+    // proof pipeline is empty)
     if (!proof_pipeline.filterContoursOutput().isEmpty()) {
+      
+      // Max amount of balls that can be seen at once
       for (int i = 0; i < 5; i++) {
+    
         try {
-          SmartDashboard.putNumber("i", i);
+          SmartDashboard.putNumber("Balls Seen", i);
+          
+          // Rect is a class that is used to find the center of the
+          // object. There is a rectangle drawn around the contours found, and
+          // then it will find the center of rectangle.
           Rect r = Imgproc.boundingRect(proof_pipeline.filterContoursOutput().get(i));
+          
+          // TODO: See if the code works without this confusing logic
           synchronized (imgLock) {
+            // Stores values in Contours
             Contours[i].CenterX = (r.x + (r.width / 2));
             Contours[i].CenterY = (r.y + (r.height / 2));
             Contours[i].Area = (r.area());
           }
-        } catch (Exception e) {
+        } 
+        
+        catch (Exception e) {
           // System.out.println("CAN'T HAVE EMPTY CONTOURS");
           System.out.println(e);
           return Contours;
         }
+
       }
+
     }
 
     // DEBUGGING
@@ -93,8 +112,6 @@ public class USBVision extends SubsystemBase {
     //       centerX = r.x + (r.width / 2);
     //     }
     // }
-
-    // DEBUGGING
     // SmartDashboard.putNumber("Center 1", centers[0]);
     // SmartDashboard.putNumber("Center 2", centers[1]);
     // SmartDashboard.putNumber("Center 3", centers[2]);
